Load module controller to turn on MainlineTestModuleController for Mainline Test Mapping.
MainlineTestModuleController is a module controller to run tests
when the mainline module is preloaded on the device, and it’s
disabled by default(reason: b/181724969#comment12)

It usually costs maintenance efforts when more tests are added
in Mainline Test Mapping and are targeting a specific platform
whose the sdk version isn't finalized.

This change provides a way to skip tests running on a specific
platform when the sdk is not finalized.

Design: go/enable-mainline-test-module-controller-in-mtm

Bug: 197914757
Test: unitttest
      presubmit
Change-Id: I54fb8da8de2bb882967e422b65feed349d0d49fa