make IP configurable for VmRemoteDevice type
VmRemoteDevice type was created to represent large instances
in GCE, however it is also suitable for large physical machines
that can run many AVDs simultaneously. The main differences are
in device creation and the presence of a pre-configured host
IP in the case of large physical machine.

This change also abstracts out the notion of having a pre-
configured IP into a dedicated interface so that checking for
it can be done with the interface instead of a list of device
classes.

Bug: 202400161
Test: existing unit tests & e2e
Change-Id: I8ad40c835a031e52183674f4c4959d076ef79d8e