Add checks around ADB file transfers for Checker tests in `ArtRunTest`.
When pulling a file from the device-under-test as part of Checker
tests in the ART run-test TradeFed test runner, perform additional
checks to verify that the file file was transferred correctly, by
comparing the sizes and MD5 digests of the original file (on device)
and its (local) copy. If one of those checks fail, report it as a
(uncaught) exception, so that TradeFed does not report it as an actual
test failure, but as a test error (i.e. caused by the test
environment, not by the test itself, or the code it is exercising).

This change establishes a separation between the origins of failed
file comparisons in Checker tests, which has the following benefits:
- errors (caused by the test environment) in Checker tests will no
  longer be reported as an genuine test failure, thus removing a
  source of test flakes;
- when a Checker test fails because of an error (caused by the test
  environment), TradeFed may decide to retry it (depending on how it
  is configured to run), before reporting it as an actual error.

Test: m tradefed tradefed-tests \
        && tools/tradefederation/core/javatests/run_tradefed_tests.sh \
             --class com.android.tradefed.testtype.ArtRunTestTest
Test: atest art-run-test-004-checker-UnsafeTest18
Bug: 218030517
Bug: 181646989
Bug: 162408889
Change-Id: Ie155b9cc5df5d6ae7374d9f0aa1b9247f9c9311d