Report Checker test environment errors as exceptions in `ArtRunTest`.
No longer encapsulate nor process these errors in instances of
`CheckerTestException` (and remove the definition of this exception
class). Instead, signal test environment errors as (uncaught)
exceptions, so that TradeFed does not report them as actual test
failures, but as a test errors (and possibly tries to run the test
again).

Test: m tradefed tradefed-tests \
        && tools/tradefederation/core/javatests/run_tradefed_tests.sh \
             --class com.android.tradefed.testtype.ArtRunTestTest
Test: atest art-run-test-004-checker-UnsafeTest18
Bug: 218030517
Bug: 181646989
Bug: 162408889
Change-Id: I71c1f22a3b9add9c26a9397795a27d5fa1fd75cd