Fix `optimize-mainline-test` option for T+ device.
Starting with T device, the preload mainline module may be located
in /data/apex/decompressed, so the previous assumption that updated
mainline module is always located in /data is not true, which caused
`optimize-mainline-test` not work for T+ device.

After confirmation, updated mainline modules are located in
/data/apex/active, /data/app-staging, or /data/apex/sessions, so the
optimization logic should only look for those locations.

Bug: 235308500
Test: unittest
      forrest
Change-Id: I00a7d77d79adf034cbff12eb1278e871301f259f