Change test module name with parameterized name for Test Mapping in Mainline
The issue only happens in Mainline when a test is pass for moduleA[a] but fails for moduleA[a+b], however, the current logic reports failures for moduleA, which confuses ATI and make moduleA as flaky test.

The solution is to treat module name with parameterized name so that ATI can differentiate them.

Verified with a forrest run: https://atp.googleplex.com/test_runs/240395043

Bug: 233563737
Bug: 220265257
Test: unitttest/forrest

Change-Id: I9edfccf3b6d6f619fd0e3a6243a73fa6724f7163