add TargetFileUtils to core tradefed
This is the last utility in VTS that is used by many tests
and prevents us from deleting vts-core-tradefed-harness.
The only other thing remaining is the kernel test runner.

I couldn't find an alternative for this functionality, so
I've moved it as-is, except I replaced use of EasyMock
with Mockito, as seems to be the style here.

The code could be simplified, but more important is avoiding
forking our test infra like this.

Bug: 274790216
Test: m tradefed-all && ./javatests/run_tradefed_tests.sh
Change-Id: Iad791e5abd889076920b761140d17f35c8b3d20e