Update shard split logic for ExecutableBaseTest
The previous shard split logic would split the ExecutableBaseTest into
all of its individual internal tests. This caused problems for
downstream logic, such as gcov kernel coverage, that rely on these
tests to be grouped as best as possible. In addition, this splitting up
into each single test was unnecessary as they end up getting grouped
again on the shards.

The new logic takes into account the number of shards and splits the
tests into that many ExecutableBaseTest tests.

Bug: 300343059
Test: ExecutableTargetTestTest#testShard_SplitPerShard, ExecutableTargetTestTest#testShard_SplitPerTestCommand

Change-Id: Ibc0f9f64524a5f5075cf0e96e86f5752e587947b
Signed-off-by: Joe Fradley <joefradley@google.com>