Fix incorrect logic for calculating retry stats.
If a module crashes the reporting may be incomplete, so there is no actual guarantee that every method is reported. This is particularly true for tests other than InstrumentationTest which do not report all the tests when a crash occurs.

Added one unit test to verify the retry metrics are calculated correctly.

Test: Unit tests
Bug: b/400683699
Change-Id: I1c0fb69ac0ce2c8b859f7debeab8a65ba33b31b3