Avoid leaking performance post-processors between test modules
Currently, ITestSuite#run() initializes moduleListeners once, with
per-module handling appending to it or wrapping it. As a result, any
post-processors added by listenerWithPostProcessorsForPerfModule() are
retained for subsequent test modules. In the case of a
MetricFilePostProcessor, this results in an additional results file for
each prior test module.

Avoid this leak by narrowing the scope of moduleListeners so it's
initialized for each module.

Bug: 404654403
Test: atest power-idle-display-on power-manual
Test: mm && ./javatests/run_tradefed_tests.sh
Change-Id: I01fb38c5d168a1f7db4bc1b5c1f191d7db269406