Buffer results for module listeners too
Tying module listeners with invocation listeners so that they receive
test results at the same time. This is done at the ITestSuite level so
moduleDefinition or other classes do not have to care about module
listeners. In order for logging to work properly, ModuleListener now
setLogSaver() so that module level postprocessors have a copy of the
log saver, which they can use to save files.

Bug: 326452860
Bug: 404936835
Test: http://ab/I17700010378898875,
https://atp.googleplex.com/test_runs/06766959431432273920, https://atp.googleplex.com/test_runs/01775108471608311808
Change-Id: I98e03a1b663b94d9038a569a91be2609d48c8f80